Initial Commit
Added some file structure to the project including bin and src.
The src directory contains 4 directories: Main, Translator, Spaces
and Objects. Within Objects direcory an ObjectEntity class was
created intended to be the base 'Object' (a thing) in the model.
Within Spaces a Space class was created intended to be the base
'Space' where 'Objects' live. Main will house all user interface
and instantiation of the model. Translator will house all scripts
which will translate language into model actions.

A makefile was added which will be the main way to build the
project. Of course a user is welcome to manually compile all the
files.